\name{Algorithm-package}
\alias{Algorithm-package}
\alias{Algorithm}
\docType{package}
\title{
\packageTitle{Algorithm}
}
\description{
\packageDescription{Algorithm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Algorithm}
\packageIndices{Algorithm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Algorithm}

Maintainer: \packageMaintainer{Algorithm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
