% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{The Euclidean Algorithm}
\usage{
euclidean(value1, value2)
}
\arguments{
\item{value1}{A numeric scalar or integer value.}

\item{value2}{A numeric scalar or integer value.}
}
\value{
the greatest common divisor of two numbers i.e value1 and value2
}
\description{
the Euclidean algorithm is an efficient method for computing the greatest common divisor of two integers, the largest number that divides them both without a remainder.
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
